feat : 관광지 저장 로직 구현

+++ BF_TRIP/BF_TRIP/Network/DTO/ResponsePlaceDTO.swift
@@ -10,7 +10,7 @@ import Foundation
 struct ResponsePlaceDTO: Identifiable, Codable, Hashable {
     
     let id = UUID().uuidString
-    let contentId: UInt64
+    let contentId: Int
     let contentTitle: String
     let addr: String
     let gpsX: Double
